trying to fix GUI mode

+++ include/utils.hpp
@@ -1,10 +1,10 @@
 #pragma once
-
+#include <stdint.h>
 // static void printBorder();
 // static void printLogo();
 // static void setBackground(int color);
 // static void debugRegs();
 // bool checkMem(char str[], int address);
-int getMemory(int Address);
+uint8_t getMemory(int Address);
 void amogus();
 void printColorPallet();
